Showing scancodes on LEDs (parallel port)
Showing scancode on external LEDs connected through parallel port.

This example sends the scancode of the key pressed to the parallel port so that it is visible on LEDs.
Port 37A is the control port. Our interest is with bit 4 of this port which enables the IRQ 7 triggering by the ACK pin. We have attached a circuit that connects 8 LEDs with the parallel port pins.