Slowly turning on the bulb by increasing power
Slowly turning on a bulb by gradually increasing the power provided.

We used the parallel port to control a slightly complicated circuit. We used the parallel port interrupt. We are using a 220 V bulb with AC input. AC current is 50Hz sine wave. We made our circuit such that it triggers the parallel port interrupt whenever the sine wave crosses zero. We have control of passing the AC current to the bulb. We control it such that in every cycle only a fixed percentage of time the current passes on to the bulb. Using this we can control the intensity or glow of the bulb.