Elementary multitasking of two threads
Elementary multitasking of two threads.

This is the basic example of multitasking. The two subroutines rotate bars by changing characters at the two corners of the screen and have infinite loops. By hooking the timer interrupt and saving and restoring the registers of the tasks one by one, it appears that both tasks are running simultaneously.